requirements.txt
NREL-pySAM==1.2.1
